requirements.txt
arabic-reshaper==3.0.0
attrs==24.2.0
Brotli==1.1.0
cffi==1.17.1
chardet==4.0.0
cssselect2==0.7.0
dj-database-url==2.2.0
Django==5.1.8
django-cors-headers==4.3.1
django-filter==24.3
django-jalali==6.2.1
django-mysql==4.14.0
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.0
fonttools==4.53.1
html5lib==1.1
ippanel==2.0.7
isodate==0.6.1
jalali_core==1.0.0
jdatetime==5.0.0
lxml==5.2.2
mysql-connector-python==8.4.0
mysql-server-has-gone-away==2.0.0
pillow==10.4.0
platformdirs==4.2.2
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
tinycss2==1.3.0
tzdata==2024.1
weasyprint==62.3
webencodings==0.5.1
zeep==4.2.1
zopfli==0.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-extensions==4.1
python-decouple==3.8
psycopg2-binary==2.9.10
aiosmtpd